requirements.txt
torch
torchvision
tqdm
pyyaml
diffusers
transformers
accelerate
safetensors
pillow
matplotlib
